+++ plugins/coredns/latency/plugin.go
@@ -11,38 +11,39 @@ import (
 
 const (
 	SLIPluginVersion = "prometheus/v1"
-	SLIPluginID      = "sloth-common/kubernetes/kooper/latency"
+	SLIPluginID      = "sloth-common/coredns/latency"
 )
 
 // We use a combination of `>0` on the division denominator (so we don't divide by 0 to avoid getting `NaN`)
 // and then `OR on vector(1)` when the `>0` took effect to return a `0` error ratio.
 var queryTpl = template.Must(template.New("").Option("missingkey=error").Parse(`
 1 - ((
-  sum(rate(kooper_controller_processed_event_duration_seconds_bucket{ {{ .filter }}controller="{{ .controller }}",le="{{ .bucket }}" }[{{"{{ .window }}"}}]))
+  sum(rate(coredns_dns_request_duration_seconds_bucket{ {{ .filterError }}le="{{ .bucket }}" }[{{"{{ .window }}"}}]))
   /
-  (sum(rate(kooper_controller_processed_event_duration_seconds_count{ {{ .filter }}controller="{{ .controller }}" }[{{"{{ .window }}"}}])) > 0)
+  (sum(rate(coredns_dns_request_duration_seconds_count{ {{ .filterTotal }} }[{{"{{ .window }}"}}])) > 0)
 ) OR on() vector(1))
 `))
 
-// SLIPlugin will return a query that will return the event handling latency by using the kooper's
-// event handling time measurements.
+// SLIPlugin will return a query that will calculate the the request handling latency in coreDNS.
 func SLIPlugin(ctx context.Context, meta, labels, options map[string]string) (string, error) {
 	bucket, err := getBucket(options)
 	if err != nil {
 		return "", fmt.Errorf(`could not get bucket: %w`, err)
 	}
 
-	controller, err := getController(options)
-	if err != nil {
-		return "", fmt.Errorf(`could not get controller: %w`, err)
+	filter := getFilter(options)
+	filterTotal := filter
+	filterError := filter
+	if filterError != "" {
+		filterError = filter + ","
 	}
 
 	// Create query.
 	var b bytes.Buffer
 	data := map[string]string{
-		"controller": controller,
-		"bucket":     bucket,
-		"filter":     getFilter(options),
+		"bucket":      bucket,
+		"filterTotal": filterTotal,
+		"filterError": filterError,
 	}
 	err = queryTpl.Execute(&b, data)
 	if err != nil {
@@ -66,21 +67,9 @@ func getBucket(options map[string]string) (string, error) {
 	return bucket, nil
 }
 
-func getController(options map[string]string) (string, error) {
-	controller, ok := options["controller"]
-	if !ok || (ok && controller == "") {
-		return "", fmt.Errorf(`"controller" option is required`)
-	}
-
-	return controller, nil
-}
-
 func getFilter(options map[string]string) string {
 	filter := options["filter"]
 	filter = strings.Trim(filter, "{},")
-	if filter != "" {
-		filter += ","
-	}
 
 	return filter
 }